% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datacleaningfunctions.R
\name{comments}
\alias{comments}
\title{Comments}
\usage{
comments(cover, fuel1000, duff, fine, saps, seeds, tree)
}
\arguments{
\item{cover}{}

\item{fuel1000}{}

\item{duff}{}

\item{fine}{}

\item{saps}{}

\item{seeds}{}

\item{tree}{}
}
\value{
A list of comments excluding some repeated ones that aren't relevant
}
\description{
The `comments` function compiles unique, non-empty comments from protocols cover, 1000 HR Fuels, Duff, Fine Fuels, Saplings, Seedlings, and Trees, excluding common placeholders. It categorizes comments by protocol, including macro plot names, monitoring statuses, and associated comments.
}
\examples{
comments(cover, fuel1000,duff, fine, saps, seeds, tree)
}
