% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AKR_datacleaningfunctions.R,
%   R/datacleaningfunctions.R, R/datacleaningfunctions_AKR.R
\name{duff_qc}
\alias{duff_qc}
\title{Duff Litter}
\usage{
duff_qc(duff)

duff_qc(duff)

duff_qc(duff)
}
\arguments{
\item{duff}{}
}
\value{
A list of flags or data issues in the duff csv

A list of flags or data issues in the duff csv

A list of flags or data issues in the duff csv
}
\description{
The duff_qc function performs quality control checks on duff data. It first ensures that all duff number transects are equal to 4, printing a message to confirm this. If any transects have a different count, it generates an error message, indicating the row and comment associated with the discrepancy. The function then groups the data by macro plot and sample event date, checking for errors in duff transect values and sample locations within each group. It reports any mismatches and provides details about the offending values and sample events. Additionally, the function checks for outlier values in litter depth and duff depth using the Rosner test, generating messages if outliers are detected. Lastly, it checks for null values in both variables, reporting the affected sample events, transects, and associated comments. The function compiles all error messages in the 'flags' vector and returns it.

The duff_qc function performs quality control checks on surface fuels duff
and litter dataset. It first validates that the number of transects sampled is
consistent with monitoring protocols (*may need editing to be consistent with
program protocols) sample events have a different count, it generates an error
message, indicating the sample event (macroplot + date), and transect numbers.
It reports any missing or duplicate sample locations. Additionally, the function
checks for outlier values in litter depth and duff depth using the Rosner test,
generating messages if outliers are detected. Lastly, it checks for null values
in both variables, reporting the affected sample events, transects, and associated
comments. The function compiles all error messages in the 'flags' vector and returns it.

The duff_qc function performs quality control checks on duff data. It first ensures that all duff number transects are equal to 4, printing a message to confirm this. If any transects have a different count, it generates an error message, indicating the row and comment associated with the discrepancy. The function then groups the data by macro plot and sample event date, checking for errors in duff transect values and sample locations within each group. It reports any mismatches and provides details about the offending values and sample events. Additionally, the function checks for outlier values in litter depth and duff depth using the Rosner test, generating messages if outliers are detected. Lastly, it checks for null values in both variables, reporting the affected sample events, transects, and associated comments. The function compiles all error messages in the 'flags' vector and returns it.
}
\examples{
duff_qc(duff)
duff_qc(duff)
duff_qc(duff)
}
