% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datacleaningfunctions.R
\name{fine_fuels_qc}
\alias{fine_fuels_qc}
\title{Fine Litter Fuels}
\usage{
fine_fuels_qc(fine)
}
\arguments{
\item{fine}{}
}
\value{
A list of flags or data issues in the fine fuels csv
}
\description{
The fine_fuels_qc function performs quality control checks on fine fuels data. It verifies if the transect lengths for One Hour, Ten Hour, and Hundred Hour are consistently 6, checks for misplaced fuel values in the transect length rows, and identifies sample events with missing transect length information. The function also groups the data by macro plot and sample event date, then checks for errors in fine transect values, including patterns that deviate from the expected range (1 to 4) and duplicated values. If all checks pass, the function validates that the fine transect values are correct. Any inconsistencies or issues detected during the checks are flagged for further investigation.
}
\examples{
fine_fuels_qc(fine)

}
