% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datacleaningfunctions.R
\name{format_flags}
\alias{format_flags}
\title{Format flags}
\usage{
format_flags(flags, samp, mtype, comments)
}
\arguments{
\item{flags}{}

\item{samp}{}

\item{mtype}{}

\item{comments}{}
}
\value{
Excel file separating issues by plot
}
\description{
The format_flags function generates an Excel output that consists of multiple sheets, each corresponding to a distinct macro plot and an additional sheet for comments. For each macro plot sheet, the data is organized with a column named "Issue" containing flagged information. Additional columns include "Resolved," "Resolved_by," "Action_need," and "Other_notes," providing a structured format for documenting and tracking the resolution status of flagged issues.
}
\examples{
format_flags(flags, samp, mtype, comments)
}
