% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AKR_datacleaningfunctions.R,
%   R/datacleaningfunctions.R, R/datacleaningfunctions_AKR.R
\name{fuel1000_qc}
\alias{fuel1000_qc}
\title{Fuel 1000}
\usage{
fuel1000_qc(fuel1000)

fuel1000_qc(fuel1000)

fuel1000_qc(fuel1000)
}
\arguments{
\item{fuel1000}{}
}
\value{
A list of flags or data issues in the fuel csv

A list of flags or data issues in the fuel csv

A list of flags or data issues in the fuel csv
}
\description{
The fuel1000_qc function conducts quality control checks on surface fuels data within the fuel1000 data set. It begins by validating decay class, transect, and slope values, and ensures that tree diameter values are within the specified range. It then identifies missing diameter values and checks for the presence of blank DIA for each unique macro plot and monitoring status combination. Additionally, it performs outlier tests using Rosner test for the "Slope" and "Dia" columns, logging any detected outliers and their details.

The fuel1000_qc function conducts quality control checks on surface fuels data
within the surface fuels coarse woody debris (1000-hr) data set. It begins by
validating decay class, transect, and slope values, and ensures that log diameter
values, transects sampled, and sampled transect length are within the specified range.
*IF ADJUSTING FOR DIFFERENT PROGRAM – you must edit code to include values specific to
your program. It also identifies missing diameter values. Additionally, it performs
outlier tests using Rosner test for the "Slope" and "Dia" columns, logging any detected
outliers and their details.

The fuel1000_qc function conducts quality control checks on surface fuels data within the fuel1000 data set. It begins by validating decay class, transect, and slope values, and ensures that tree diameter values are within the specified range. It then identifies missing diameter values and checks for the presence of blank DIA for each unique macro plot and monitoring status combination. Additionally, it performs outlier tests using Rosner test for the "Slope" and "Dia" columns, logging any detected outliers and their details.
}
\examples{
fuel1000_qc(fuel1000)
fuel1000_qc(fuel1000)
fuel1000_qc(fuel1000)
}
