% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datacleaningfunctions.R
\name{pbsev_qc}
\alias{pbsev_qc}
\title{Post Burn Severity Substrate Checks}
\usage{
pbsev_qc(pbsev)
}
\arguments{
\item{pbsev}{}
}
\value{
A list of flags or data issues with post burn severity substrate data in the post burn severity csv
}
\description{
This function checks for incorrect values for NumTran, TranLen, NumPtsTran, substrate and vegetation values
in the post burn severity data frame. Values may need to be edited according to park specific protocol. The
default values are used in Saguaro's monitoring program. The function also checks that there are
4 transects for each macroplot with 10 sample points per transect. Any errors are flagged with information
about the problem sample event and macroplot.
}
\examples{
pbsev_qc(tree)
}
