% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datacleaningfunctions.R
\name{sample_event_qc}
\alias{sample_event_qc}
\title{Sample event quality control checks}
\usage{
sample_event_qc(samp, mtype)
}
\arguments{
\item{mtype}{}
}
\value{
A list of flags or data issues in the sample events csv
}
\description{
The sample_event_qc function is designed to perform quality control checks on sample event data associated with different vegetation types.
It takes two arguments, samples (a dataset containing sample event records) and mtype (the vegetation type to be checked).
The function first filters the dataset to include only the specified vegetation type.
It then checks for missing values (NAs) in the MonitoringStatus_Name, Protocols, and Visited columns.
If any NAs are detected, they are added to a list of flags.
Additionally, the function checks for mislabeled or missing monitoring statuses, protocols, and visited values.
}
\examples{
sample_event_qc(samp, "PSME")

}
