% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AKR_datacleaningfunctions.R,
%   R/datacleaningfunctions.R, R/datacleaningfunctions_AKR.R
\name{saplings_qc}
\alias{saplings_qc}
\title{Saplings}
\usage{
saplings_qc(saps)

saplings_qc(saps)

saplings_qc(saps)
}
\arguments{
\item{saps}{}
}
\value{
A list of flags or data issues in the saplings csv

A list of flags or data issues in the saplings csv

A list of flags or data issues in the saplings csv
}
\description{
The saplings_qc function is designed to perform quality control checks on a dataset of saplings. It specifically focuses on detecting and addressing potential errors related to the macroplot size recorded for each sapling. The function begins by checking whether the macroplot size is consistently recorded as 0 for all saplings, which would indicate that they were not sampled by diameter class. If the unique value of macroplot size is found to be 0, the function prints "TRUE" to indicate the absence of errors. In case the macroplot size is not uniformly 0, the function generates an error message indicating that the value(s) should be replaced with 0 in the macroplot size list for saplings.

The saplings_qc function is designed to perform quality control checks on a
dataset of saplings. It specifically focuses on detecting and addressing potential
errors related to the macroplot size recorded for each sapling. The function begins
by checking whether the macroplot size is consistently recorded as 0 for all saplings,
which would indicate that they were not sampled by diameter class. If the unique
value of macroplot size is found to be 0, the function prints "TRUE" to indicate
the absence of errors. In case the macroplot size is not uniformly 0, the function
 generates an error message indicating that the value(s) should be replaced with
  0 in the macroplot size list for saplings.

The saplings_qc function is designed to perform quality control checks on a dataset of saplings. It specifically focuses on detecting and addressing potential errors related to the macroplot size recorded for each sapling. The function begins by checking whether the macroplot size is consistently recorded as 0 for all saplings, which would indicate that they were not sampled by diameter class. If the unique value of macroplot size is found to be 0, the function prints "TRUE" to indicate the absence of errors. In case the macroplot size is not uniformly 0, the function generates an error message indicating that the value(s) should be replaced with 0 in the macroplot size list for saplings.
}
\examples{
saplings_qc(saps)

saplings_qc(saps)

saplings_qc(saps)

}
