% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datacleaningfunctions.R
\name{tree_CBH_qc}
\alias{tree_CBH_qc}
\title{Tree CBH and Tree CBH Range}
\usage{
tree_CBH_qc(tree)
}
\arguments{
\item{tree}{}
}
\value{
A list of flags or data issues with CBH in the tree csv
}
\description{
The tree_CBH_qc function performs quality control checks on the 'tree' dataset. It ensures that dead trees have blank live crown base height values and flags any discrepancies. For living trees, it checks for unreasonable live crown base height values using the Rosner test and flags outliers. The function also verifies the relationship between ladder base height (LaddBaseHt) and ladder max height (LaddMaxHt), ensuring the former is always less than the latter. It checks that entries with LaddBaseHt values have the correct species symbol ('CANOPY'), status ('L'), SubFrac (1000), and TagNo (0), flagging any inconsistencies and providing detailed information about problematic rows.
}
\examples{
tree_CBH_qc(tree)
}
