% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AKR_datacleaningfunctions.R,
%   R/datacleaningfunctions.R, R/datacleaningfunctions_AKR.R
\name{tree_crown_class_qc}
\alias{tree_crown_class_qc}
\title{Crown Class}
\usage{
tree_crown_class_qc(tree)

tree_crown_class_qc(tree)

tree_crown_class_qc(tree)
}
\arguments{
\item{tree}{}
}
\value{
A list of flags or data issues with crown class in the tree csv

A list of flags or data issues with crown class in the tree csv

A list of flags or data issues with crown class in the tree csv
}
\description{
The tree_crown_class_qc function classifies trees based on their diameter at breast height (DBH) and validates the correctness of crown class values for different tree statuses. It ensures that for trees with a blank or dead status and blank DBH, the crown class is either "BBD," "CUS," "DD," or blank. In the case of live trees with blank DBH, it verifies that the crown class is blank. For overstory trees with a dead status, the function checks for acceptable crown class values such as "BAD," "CS," "LBS," "RS," or blank. Similarly, for live overstory trees, it ensures crown class values are within categories "C," "D," "I," "O," "SC," or blank. Additionally, for pole trees with a dead status, the function checks for valid crown class values like "BBD," "CUS," "DD," "X," or blank. The function flags instances where crown class values deviate from these specified categories, providing quality control for tree data.

The tree_crown_class_qc function classifies trees based on their diameter at
breast height (DBH) and validates the correctness of crown class values for
different tree statuses. It ensures that for trees with a blank or dead status
and blank DBH, the crown class is either "BBD," "CUS," "DD," or blank. In the
case of live trees with blank DBH, it verifies that the crown class is blank.
For overstory trees with a dead status, the function checks for acceptable crown
class values such as "BAD," "CS," "LBS," "RS," or blank. Similarly, for live
overstory trees, it ensures crown class values are within categories
"C," "D," "I," "O," "SC," or blank. Additionally, for pole trees with a dead status,
the function checks for valid crown class values like "BBD," "CUS," "DD," "X," or blank.
The function flags instances where crown class values deviate from these specified
categories, providing quality control for tree data.

The tree_crown_class_qc function classifies trees based on their diameter at breast height (DBH) and validates the correctness of crown class values for different tree statuses. It ensures that for trees with a blank or dead status and blank DBH, the crown class is either "BBD," "CUS," "DD," or blank. In the case of live trees with blank DBH, it verifies that the crown class is blank. For overstory trees with a dead status, the function checks for acceptable crown class values such as "BAD," "CS," "LBS," "RS," or blank. Similarly, for live overstory trees, it ensures crown class values are within categories "C," "D," "I," "O," "SC," or blank. Additionally, for pole trees with a dead status, the function checks for valid crown class values like "BBD," "CUS," "DD," "X," or blank. The function flags instances where crown class values deviate from these specified categories, providing quality control for tree data.
}
\examples{
tree_crown_class_qc(tree)
tree_crown_class_qc(tree)
tree_crown_class_qc(tree)
}
