% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datacleaningfunctions.R
\name{tree_damage_qc}
\alias{tree_damage_qc}
\title{Tree damage}
\usage{
tree_damage_qc(tree)
}
\arguments{
\item{tree}{}
}
\value{
A list of flags or data issues with damage class in the tree csv
}
\description{
The `tree_damage_qc` function validates the accuracy of tree damage codes in a dataset. It checks if the codes provided (DamCd1, DamCd2, DamCd3, DamCd4, DamCd5) are within the predefined acceptable list of damage codes. The function returns "TRUE" if all codes are correct and included in the acceptable list, and "FALSE" with details of any discrepancies found. The acceptable list includes codes such as "ABGR," "BIRD," "BLIG," and others. If discrepancies are detected, the function adds corresponding entries to the quality control flags, specifying the MacroPlot.Name and Monitoring.Status for the rows where the issues occur.
}
\examples{
tree_damage_qc(tree)
}
