% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datacleaningfunctions.R
\name{tree_dbh_qc}
\alias{tree_dbh_qc}
\title{Tree DBH}
\usage{
tree_dbh_qc(tree)
}
\arguments{
\item{tree}{}
}
\value{
A list of flags or data issues with DBH in the tree csv
}
\description{
The `tree_dbh_qc` function conducts quality control on the Diameter at Breast Height (DBH) values in a tree dataset. It first checks if all DBH values are under 2.5; if any exceed this limit, it flags the events with the specific DBH values, MacroPlot.Name, and Monitoring.Status. Subsequently, the function performs a check for outlier values using a rosner test on the DBH values. If outliers are detected, it adds entries to the quality control flags with details on the outlier values, the corresponding events (MacroPlot.Name and Monitoring.Status), and statistics on the maximum, minimum, and mean of the DBH values in the dataset.
}
\examples{
tree_dbh_qc(tree)
}
