% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datacleaningfunctions.R
\name{tree_dead_to_alive_DBH_change_qc}
\alias{tree_dead_to_alive_DBH_change_qc}
\title{Tree dead to alive and changing dbh}
\usage{
tree_dead_to_alive_DBH_change_qc(tree)
}
\arguments{
\item{tree}{}
}
\value{
A list of flags or data issues with trees resurrecting and changing dbh when dead in the tree csv
}
\description{
The tree_dead_to_alive_DBH_change_qc function examines tree data for instances of dynamic status changes and potential Diameter at Breast Height (DBH) alterations across multiple sampling events. It systematically identifies trees that consistently exhibit DBH changes despite being recorded as dead. Additionally, the function captures cases where trees switch between alive and dead statuses, providing specific information on dates and circumstances. Flagged issues are reported in the 'flags' variable, separating DBH-related concerns and occurrences of trees transitioning from dead to alive.
}
\examples{
tree_dead_to_alive_DBH_change_qc(tree)
}
