% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datacleaningfunctions.R
\name{tree_header_qc}
\alias{tree_header_qc}
\title{Tree header}
\usage{
tree_header_qc(tree)
}
\arguments{
\item{tree}{}
}
\value{
A list of flags or data issues with header in the tree csv
}
\description{
The `tree_header_qc` function performs quality control checks on the header information of the tree data. It first ensures that all macroplot sizes (`MacroPlotSize`) are either 0.1 or blank, with detailed information about problematic events stored in the `flags` variable. Similarly, it verifies that all snagplot sizes (`SnagPlotSize`) are 0.1 or blank. The function also checks that break point diameters (`BrkPntDia`) are either 15.1 or blank. For each of these checks, it provides a logical value and details about any non-compliance events stored in the `flags` variable, which is returned by the function.
}
\examples{
tree_header_qc(tree)
}
