% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datacleaningfunctions.R
\name{tree_height_qc}
\alias{tree_height_qc}
\title{Tree height}
\usage{
tree_height_qc(tree)
}
\arguments{
\item{tree}{}
}
\value{
A list of flags or data issues with height in the tree csv
}
\description{
The `tree_height_qc` function categorizes trees based on diameter at breast height (DBH) into pole, overstory, and blank trees. For overstory trees, it conducts a Rosner test to identify height outliers, reporting details if four or more valid samples are available. The function also checks pole tree data to ensure entered height classes fall within the acceptable range of 1 to 10 or are blank, flagging any deviations and providing event details in the `flags` variable returned by the function.
}
\examples{
tree_height_qc(tree)
}
