% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datacleaningfunctions.R
\name{tree_sample_area_qc}
\alias{tree_sample_area_qc}
\title{Tree sample area}
\usage{
tree_sample_area_qc(tree)
}
\arguments{
\item{tree}{}
}
\value{
A list of flags or data issues with header in the tree csv
}
\description{
The `tree_sample_area_qc` function performs quality control checks on the sample area
information of the tree data. It first ensures that all macroplot sizes (`MacroPlotSize`)
are 0.1, with detailed information about problematic events stored in the `flags` variable.
Similarly, it verifies that all snagplot sizes (`SnagPlotSize`) are 0.1. The function also
checks that break point diameters (`BrkPntDia`) are 15.1. For each of these checks, it
provides a logical value and details about any non-compliance events stored in the `flags`
variable, which is returned by the function.
}
\examples{
tree_sample_area_qc(tree)
}
