% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datacleaningfunctions.R
\name{tree_status_lifeform_qc}
\alias{tree_status_lifeform_qc}
\title{Tree status and life form}
\usage{
tree_status_lifeform_qc(tree)
}
\arguments{
\item{tree}{}
}
\value{
A list of flags or data issues with tree status and life form in the tree csv
}
\description{
The tree_status_lifeform_qc function assesses the quality of data related to tree status and life form in the provided tree dataset. It ensures that tree statuses are either "L" or "D" and that all preferred life forms are "Tree." The function reports any deviations from these criteria, flagging problematic events and providing details such as macroplot names, monitoring statuses, tag numbers, and comments.
}
\examples{
tree_status_lifeform_qc(tree)
}
