% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datacleaningfunctions.R
\name{cover_qc}
\alias{cover_qc}
\title{Cover - species composition (metric) checks}
\usage{
cover_qc(cover)
}
\arguments{
\item{cover}{}
}
\value{
A list of flags or data issues in the cover csv
}
\description{
Cover - species composition (metric) checks
}
\examples{
cover_qc(cover)
}
