% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datacleaningfunctions.R
\name{cover_qc}
\alias{cover_qc}
\title{Cover - species composition (metric) checks}
\usage{
cover_qc(cover)
}
\arguments{
\item{cover}{}
}
\value{
A list of flags or data issues in the cover csv
}
\description{
The cover_qc function focuses on quality control checks for canopy data, particularly canopy cover class values. It takes a dataset cover as input and assesses the presence and correctness of canopy data. The function first filters the dataset to extract canopy data and then checks if all quarters of data are present for each unique macro plot and monitoring status combination. If any quarter is missing, the function logs an error message. Additionally, it validates the recorded canopy cover class values and provides feedback on any discrepancies.
}
\examples{
cover_qc(cover)
}
