% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datacleaningfunctions.R
\name{cover_uvs_qc}
\alias{cover_uvs_qc}
\title{Cover class and invasive species data}
\usage{
cover_uvs_qc(cover)
}
\arguments{
\item{cover}{}
}
\value{
A list of flags or data issues in the cover csv
}
\description{
The cover_uvs_qc function is responsible for quality control checks on cover class data, including canopy cover class and invasive species cover class values, as well as UV descriptions. It takes a dataset cover as input and conducts validation. The function first subsets the data based on UV1 values and performs checks on canopy cover class values, followed by invasive species cover class values if available. It then assesses the accuracy of UV descriptions (Protocol Name, Plot Quarter, Count Code) associated with the cover classes.
}
\examples{
cover_uvs_qc(cover)
}
