% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datacleaningfunctions.R
\name{sample_event_qc}
\alias{sample_event_qc}
\title{Sample event quality control checks
This function requires a csv file exported from FFI with the list of sample events.
The second argument is mtype which is the filter on the sample events (monitoring type).}
\usage{
sample_event_qc(samp, mtype)
}
\arguments{
\item{mtype}{}
}
\value{
A list of flags or data issues in the sample events csv
}
\description{
Sample event quality control checks
This function requires a csv file exported from FFI with the list of sample events.
The second argument is mtype which is the filter on the sample events (monitoring type).
}
\examples{
sample_event_qc(samp, "PSME")

}
