% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AKR_datacleaningfunctions.R,
%   R/datacleaningfunctions.R, R/datacleaningfunctions_AKR.R
\name{sample_event_qc}
\alias{sample_event_qc}
\title{Sample event quality control checks}
\usage{
sample_event_qc(samp, mtype)

sample_event_qc(samp, mtype)

sample_event_qc(samp, mtype)
}
\arguments{
\item{samp}{}

\item{mtype}{}
}
\value{
A list of flags or data issues in the sample events csv

A list of flags or data issues in the sample events csv

A list of flags or data issues in the sample events csv
}
\description{
The sample_event_qc function is designed to perform quality control checks on sample event data associated with different vegetation types.
It takes two arguments, samples (a dataset containing sample event records) and mtype (the vegetation type to be checked).
The function first filters the dataset to include only the specified vegetation type.
It then checks for missing values (NAs) in the MonitoringStatus_Name, Protocols, and Visited columns.
If any NAs are detected, they are added to a list of flags.
Additionally, the function checks for mislabeled or missing monitoring statuses, protocols, and visited values.

The sample_event_qc function is designed to perform quality control
checks on sample event data associated with different project units.
It takes two arguments, samples (a dataset containing sample event records)
and mtype (Project Unit Name). The function first filters the dataset to
include only the specified project unit. It then checks for missing values
(NAs) in the MonitoringStatus_Name, Protocols, and Visited columns. If any
NAs are detected, they are added to a list of flags. Additionally, the function
checks for mislabeled or missing monitoring statuses, protocols, and visited
values which will be added to flags.*IF ADJUSTING FOR DIFFERENT PROGRAM – you
must edit code to include monitoring statuses and protocols specific to your
program. This function will produce a visualization that shows which protocols
were collected each year and the years a fire occurred.

The sample_event_qc function is designed to perform quality control checks on sample event data associated with different vegetation types.
It takes two arguments, samples (a dataset containing sample event records) and mtype (the vegetation type to be checked).
The function first filters the dataset to include only the specified vegetation type.
It then checks for missing values (NAs) in the MonitoringStatus_Name, Protocols, and Visited columns.
If any NAs are detected, they are added to a list of flags.
Additionally, the function checks for mislabeled or missing monitoring statuses, protocols, and visited values.
}
\examples{
sample_event_qc(samp, "PSME")

sample_event_qc(samp, "PSME")

sample_event_qc(samp, "PSME")

}
