% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AKR_datacleaningfunctions.R,
%   R/datacleaningfunctions.R, R/datacleaningfunctions_AKR.R
\name{transect_slope_azimuth_qc}
\alias{transect_slope_azimuth_qc}
\title{Fuel fine vs 1000 hr slope azimuth consistency check}
\usage{
transect_slope_azimuth_qc(fuel1000, fine)

transect_slope_azimuth_qc(fuel1000, fine)

transect_slope_azimuth_qc(fuel1000, fine)
}
\arguments{
\item{fuel1000}{}

\item{fine}{}
}
\value{
A list of flags or data issues with slope consistency

A list of flags or data issues with slope consistency

A list of flags or data issues with slope consistency
}
\description{
The transect_slope_azimuth_qc function conducts quality control checks on slope and azimuth data in the 'fuel1000' and 'fine' datasets. It uses macroplot names in the two datasets and compares slope and azimuth values within each transect. The function flags errors if multiple different values are detected for slope or azimuth within transects. It also checks for overall consistency in slope and azimuth values across different transects.

The transect_slope_azimuth_qc function conducts quality control checks on slope
 and azimuth data in the surface fuels coarse woody debris (1000-hr) and fine woody
 debris datasets. It uses macroplot names in the two datasets and compares slope and
 azimuth values within each transect. The function flags errors if ifferent values are
 detected for slope or azimuth within transects. It also checks for overall consistency
 in slope and azimuth values across years.

The transect_slope_azimuth_qc function conducts quality control checks on slope and azimuth data in the 'fuel1000' and 'fine' datasets. It uses macroplot names in the two datasets and compares slope and azimuth values within each transect. The function flags errors if multiple different values are detected for slope or azimuth within transects. It also checks for overall consistency in slope and azimuth values across different transects.
}
\examples{
transect_slope_azimuth_qc(fuel1000, fine)
transect_slope_azimuth_qc(fuel1000, fine)
transect_slope_azimuth_qc(fuel1000, fine)
}
