% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datacleaningfunctions.R
\name{tree_dd_qc}
\alias{tree_dd_qc}
\title{Tree dead and down}
\usage{
tree_dd_qc(tree)
}
\arguments{
\item{tree}{}
}
\value{
A list of flags or data issues with dead and down trees in the tree csv
}
\description{
The `tree_dd_qc` function focuses on quality control for fallen trees in a dataset. It identifies fallen trees based on specific crown classes (BBD, CUS, DD) and creates a subset named `fallen_trees`. The function then checks if values for Diameter at Breast Height (DBH), tree height (Ht), and live crown base height (LiCrBHt) have been deleted for trees that are no longer standing. For each parameter, it reports "TRUE" if the values have been deleted as expected. If any values are still present, it flags the events with details on the specific parameter values, crown class, and associated events (MacroPlot.Name and Monitoring.Status). The quality control results are stored in the `flags` variable and returned by the function.
}
\examples{
tree_dd_qc(tree)
}
