% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AKR_datacleaningfunctions.R,
%   R/datacleaningfunctions.R, R/datacleaningfunctions_AKR.R
\name{tree_severity_qc}
\alias{tree_severity_qc}
\title{Tree non-post severity}
\usage{
tree_severity_qc(tree)

tree_severity_qc(tree)

tree_severity_qc(tree)
}
\arguments{
\item{tree}{}
}
\value{
A list of flags or data issues with severity in the tree csv

A list of flags or data issues with severity in the tree csv

A list of flags or data issues with severity in the tree csv
}
\description{
The `tree_severity_qc` function focuses on pre-burn trees by extracting relevant data based on monitoring status. It checks if all characteristic heights, scorch heights, and scorch percentage heights are blank (NA) for these pre-burn trees, flagging any discrepancies and providing event details in the `flags` variable returned by the function. If the unique values for these attributes are exclusively NA, the function returns "TRUE" for each check, indicating conformity. Otherwise, it reports the issues, listing problematic events with associated macroplot names and monitoring statuses in the `flags`.

The `tree_severity_qc` function checks if all char heights, scorch heights,
and scorch percentage heights are blank (NA) for pre-burn trees, flagging any
discrepancies and providing event details in the `flags` variable returned by the function. The function
also checks char, scorch and scorch percentage data for the sample events immediately post burn. The user
has the option to filter out pole trees for char data checks (some protocols record this data
for pole trees and some do not) with the argument filterpoles="Y" or filterpoles="N" (default).
The function checks that char height and scorch height are not blank or excessively high for
standing trees post burn, that scorch percentage is not over 100 or blank for live trees, and that trees
with crown class DD, BBD, or CUS have blank char and scorch heights and scorch percentages of
blank or 100. Errors are flagged with details about the problem sample event.

The `tree_severity_qc` function focuses on pre-burn trees by extracting relevant data based on monitoring status. It checks if all characteristic heights, scorch heights, and scorch percentage heights are blank (NA) for these pre-burn trees, flagging any discrepancies and providing event details in the `flags` variable returned by the function. If the unique values for these attributes are exclusively NA, the function returns "TRUE" for each check, indicating conformity. Otherwise, it reports the issues, listing problematic events with associated macroplot names and monitoring statuses in the `flags`.
}
\examples{
tree_severity_qc(tree)
tree_severity_qc(tree, filterpoles="Y")
tree_severity_qc(tree)
}
